added test SyncNetworkStreamClient, Perl helper script. NetworkStream fixes
We have written a Perl script which wil be internally called by the
gtest testing C++ source code as std::system(). The Perl server is on a
detached thread. Since the server dies once there an initial connection,
the Perl script first reads the input from the user, checks against the
one with a 'constant' hardcoaded value and sends a response(again, a
hardcoaded 'constant'). And on the C++ gtest source file, read the
response and ASSERT_TEST the response and buffer size. Since the Perl
server dies once after 'read' and 'write'(aka it does not go into an
infinite loop where it tries to accept other connections), we can re-use
that port for other TESTs