CMakeLists edit cleanup, wolfSSL depent added to vendor.
Removed the dummy main.cpp file from the project root and edited the
CMakeLists file which when creation of the project, we added it so that
we can link the library with the main.cpp. But it looks like we don't
need any dummy main.cpp file to act for the purpose. We need to look
more into how add_subdirectory can help with the stuff.