blueth(#17): AsyncEventLoop interface impl with epoll and tests
It took a while, but we'd implemented it successfully. The issue was
that we are not having a return statement on the function where we get a
std::shared_ptr to the event-loop implementation. It's been solved,
anyway. We'd need to consider all warnings as errors in future with a
GCC flag, it's better for sanitazation of the codebase.